requirements.txt
# Core
flask>=3.0.0
flask-cors>=4.0.0
flask-socketio>=5.3.0
gunicorn>=21.2.0

# AI Analysis
anthropic>=0.40.0

# Log Collection
watchdog>=4.0.0
scapy>=2.5.0

# Detection & Analysis
numpy>=1.26.0
scipy>=1.12.0
pyyaml>=6.0.1

# Threat Intelligence
requests>=2.31.0
feedparser>=6.0.11

# Database
peewee>=3.17.0

# Networking
gevent>=24.2.1
gevent-websocket>=0.10.1

# Utilities
python-dotenv>=1.0.0
python-dateutil>=2.8.2
ipaddress>=1.0.23
schedule>=1.2.1
jinja2>=3.1.3
